requirements.txt
beautifulsoup4
dotenv
google-generativeai
requests
sentence-transformers
torch
transformers
